Lab03-Full/requirements.txt
mysql-connector-python==9.0.0
pandas==2.2.2

